Refactor partition tracking into its own service
